Create day 2-2Q transform expression
Character.isLetterOrDigit()-This function returns a boolean value. The boolean value is true if the character is a letter or digit else is false.
If the scanned character is an operand, add it to output. 
If the scanned character is an '(', push it to the stack.
If the scanned character is an ')', pop and output from the stack until an '(' is encountered.